新增選擇電影影院日期時間按鈕
