--- cart/src/CartContext.js
@@ -1,3 +1,0 @@
-import { createContext } from "react";
-
-export const CartContext = createContext()
