+++ src/ShoppingCard/QuantityBtn.css
@@ -1,0 +1,37 @@
+/* 按钮样式 */
+.button {
+    display: inline-block; /* 使按钮可以设置宽度和内边距 */
+    padding: 10px 20px; /* 按钮内边距 */
+    font-size: 16px; /* 字体大小 */
+    font-weight: bold; /* 字体加粗 */
+    color: #fff; /* 字体颜色 */
+    background-color: #007bff; /* 背景颜色 */
+    border: none; /* 去掉边框 */
+    border-radius: 4px; /* 圆角 */
+    text-align: center; /* 文本居中 */
+    cursor: pointer; /* 鼠标悬停时显示为手型 */
+    text-decoration: none; /* 去掉下划线 */
+    transition: background-color 0.3s ease; /* 背景颜色过渡效果 */
+}
+
+.button:hover {
+    background-color: #0056b3; /* 鼠标悬停时背景颜色 */
+}
+
+.button:active {
+    background-color: #004080; /* 按钮被点击时背景颜色 */
+}
+
+/* 控制按钮容器样式 */
+.quantityBtnContainer {
+    display: flex;
+    align-items: center;
+    gap: 10px; /* 按钮之间的间距 */
+}
+
+/* 数量控制样式 */
+.quantityControl {
+    display: flex;
+    align-items: center;
+    gap: 10px; /* 按钮之间的间距 */
+}
